[#1] refactor: 응용 계층 mapper 패키지 이동
