+++ .gitignore
@@ -35,5 +35,4 @@ out/
 
 ### VS Code ###
 .vscode/
-/src/main/generated/
-/src/main/generated/
+/src/main/generated/**
